Web: Move resource specs into resource directory
