+++ web/packages/teleport/src/Discover/SelectResource/resources/index.ts
@@ -1,6 +1,6 @@
 /**
  * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
+ * Copyright (C) 2025  Gravitational, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -16,10 +16,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { SelectResource } from './SelectResource';
-export {
-  getResourcePretitle,
-  getDatabaseProtocol,
-  getDefaultDatabasePort,
-} from './resources';
-export * from './types';
+export * from './databases';
+export * from './resources';
+export * from './resourcesE';
